requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
tabulate>=0.9
python-docx>=1.0
fpdf2>=2.7
openpyxl>=3.1
openai>=1.44
python-dotenv>=1.0
pydantic>=2.0
rich>=13.0
